refactor: (#61) 전체 게시글 목록 Mock 데이터 코드를 파일로 분리
